requirements.txt
# =====================================
# Required Packages
# =====================================

# Task scheduling
APScheduler
SQLAlchemy
alembic
psycopg
pytest-postgresql
tzlocal

# =====================================
# GUI
# =====================================

PyGObject

# Browser automation
playwright

# LLM Providers
# These are required for accessing various language model APIs.
# You can install only the ones you need.
openai
mistralai
anthropic
huggingface_hub
google.generativeai
xai_sdk

# Speech Services Providers
google-cloud-texttospeech
google-cloud-speech

# for Config
pyyaml
jsonschema
platformdirs
# appdirs can be installed as an optional fallback if needed

# for Messaging
aiokafka
msgpack
prometheus_client
PyJWT

# for Audio
soundfile
sounddevice
ffmpeg
noisereduce
pydub

# For data manipulation
numpy

# ======================================
# Document parsing dependencies
# ======================================

pdfminer.six
python-docx
beautifulsoup4
Pillow
pytesseract
markdown

# =====================================
# Visualization tools
# =====================================

matplotlib

# For environment variables
python-dotenv

# HTTP client & server for API integrations
requests
aiohttp
fastapi
uvicorn[standard]

# =====================================
# Calendar backend dependencies
# =====================================

# iCalendar file parsing
icalendar

# CalDAV server support (Nextcloud, Fastmail, iCloud, etc.)
caldav>=1.3.0

# Google Calendar API
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Microsoft Outlook/365 Calendar via MSAL
msal


# Timezone handling utilities
pytz

# psutil for system monitoring and resource management
psutil>=5.9.5

# Tenacity for implementing retries and robust error handling
tenacity>=8.2.3

# Vector store providers (optional)
chromadb
faiss-cpu
pinecone-client

# =====================================
# Media/Image generation providers
# =====================================

# HTTP client for API-based providers (Stability, BlackForestLabs, XAI)
httpx

# FalAI image generation
fal-client

# Google Vertex AI / Imagen
google-cloud-aiplatform
vertexai

# Replicate API (alternative backend)
replicate

# =====================================
# RAG enhancements
# =====================================

# Hybrid search (BM25 lexical retrieval)
rank-bm25

# Observability (optional - install for production monitoring)
# opentelemetry-sdk
# opentelemetry-exporter-otlp
# opentelemetry-instrumentation-aiohttp

# =====================================
# Optional accelerator/fine-tuning dependencies
# =====================================
# See requirements-accelerators.txt for Torch, local Whisper STT, and Hugging Face fine-tuning extras.
